Create Python BLE GUI
Switch from C to Python for increased ease-of-use and
documentation. Create GUI that reads BLE from On Body device and
plots signals. It's working, but it's extremely slow right now.
We need to create packets that contain all 3 signals (mother and
baby voltages plus skin temp) and subscribe to BLE notifications
on the Base Unit to speed things up.