Speed up and match data displayed to x-axis
This code works perfectly. The framerate of the plot updates is
still slow (maybe .75 seconds per frame). The Pi appears to be
hitting 100% CPU utilization, causing it to drop some of the baby
ECG peaks and therefore miscalculate the baby's heart rate. That
said, the data displayed scrolls across the screen in about 10
seconds as expected, and the framerate is perfectly usable (just
not visually appealing). All that's left now is to add storage
capability that saves the ECG data to a USB drive. And maybe try
to iron out the issue where the program doesn't always open in
fullscreen.